the final part of the project

+++ admin/delete-category.php
@@ -6,14 +6,17 @@ if(isset($_GET['id'])) {
 
     // FOR LATER
     // update category id of posts that belong to this category to id of uncategorized category
+    $update_query = "UPDATE posts SET category_id=13 WHERE category_id=$id";
+    $update_query = mysqli_query($connection, $update_query);
 
+    if(!mysqli_errno($connection)) {
+        
+        // delete category
+        $query = "DELETE FROM categories WHERE id=$id LIMIT 1";
+        $result = mysqli_query($connection, $query);
+        $_SESSION['delete-category-success'] = "Category deleted successfully";
+    }
 
-
-
-    // delete category
-    $query = "DELETE FROM categories WHERE id=$id LIMIT 1";
-    $result = mysqli_query($connection, $query);
-    $_SESSION['delete-category-success'] = "Category deleted successfully";
 }
 header('location: ' . ROOT_URL . 'admin/manage-categories.php');
 die();
